+++ src/web/containers/MissionControllerPanel/MissionSpeedSettings/MissionSpeedSettings.less
