+++ src/web/containers/MissionControllerPanel/MissionSpeedSettings/MissionSpeedSettings.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { Speeds } from "../../shared/JAIAProtobuf";
-import { GlobalSettings, Save } from "../../missions/settings";
+import { Speeds } from "../../../shared/JAIAProtobuf";
+import { GlobalSettings, Save } from "../../../missions/settings";
 
 import Slider from "@mui/material/Slider";
 import { amber } from "@mui/material/colors";
